Import the whole damned 11.3 usb source subtree into 12.0 to try to fix the Vx1190 problem that we still don't really understand.

+++ main/usb/vmusb/daqconfig/CV1495sc.cpp
@@ -247,7 +247,7 @@ CV1495sc::Initialize(CVMUSB& controller)
 
   // Execute the list:
 
-  int16_t buffer;
+  int32_t buffer;
   size_t  bytesread;
   int status  = controller.executeList(list, &buffer, sizeof(buffer), &bytesread);
